requirements-dev.txt
-e .
aiodns==2.0.0
aiofiles==0.4.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.2.0
cffi==1.12.3
chardet==3.0.4
idna==2.8
multidict==4.5.2
pycares==3.0.0
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
slackclient[optional]==2.2.1
yarl==1.3.0
